// lib/api/generated/schemas/invoiceDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */
import type { InvoiceDtoPlan } from './invoiceDtoPlan';
import type { InvoiceDtoStatus } from './invoiceDtoStatus';

export interface InvoiceDto {
  /** Номер счёта в системе Click */
  invoiceId: number;
  /** Наш номер заказа */
  orderId: number;
  plan: InvoiceDtoPlan;
  amountUzs: number;
  /** Куда ушёл счёт */
  phone: string;
  status: InvoiceDtoStatus;
}